- Added unit tests for update
